requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
timm==0.6.7
fastai==2.7.9
numpy==1.23.1
python-multipart==0.0.5
uvicorn==0.17.6
pillow==9.1.0
Flask==2.1.2
Flask-RESTful==0.3.9
boto3==1.22.9
Werkzeug==2.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.36
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
PyMySQL==1.0.2
python-dotenv==0.20.0
python-decouple==3.6








